chore: Fix IP address retrieval in server.js
The code changes fix the retrieval of the IP address in the `server.js` file. The previous code was using the wrong property to access the IP address from the socket object. This commit updates the code to use the correct property `socket.client.conn.request.headers` to retrieve the IP address.

Note: This message assumes that the recent user commits and recent repository commits are not relevant to the current code changes.